Add more tests; fix floating-point bugs

+++ tests/python/util.py
@@ -25,4 +25,4 @@ def load_txt(filename):
     with open(filename, 'r') as f:
         for line in f:
             content.append(float(line))
-    return np.array(content)
+    return np.array(content, dtype=np.float32)
